remove duplicationin config source in init file, preppend api version to user info routes in home route and restructure app to have models and routes files in respective folders and accordingly adjust import paths in all affected files
