+++ app/__init__.py
@@ -1,10 +1,9 @@
-from flask_api import FlaskAPI
+from flask import Flask
 from config.config import app_config
 
 
 def create_app(config_name):
-    app = FlaskAPI(__name__, instance_relative_config=False)
-    app.config.from_object('config')
+    app = Flask(__name__, instance_relative_config=False)
     app.config.from_object(app_config[config_name])
 
     return app
